Refactor Hooks and Queries
