+++ app/javascript/components/admin/ServerRecordings.jsx
@@ -6,7 +6,7 @@ import {
 import AdminNavSideBar from './shared/AdminNavSideBar';
 import SearchBarQuery from '../shared/SearchBarQuery';
 import RecordingsList from '../recordings/RecordingsList';
-import useServerRecordings from '../../hooks/queries/admin/server-recordings/useServerRecordings';
+import useServerRecordings from '../../hooks/queries/admin/server_recordings/useServerRecordings';
 import ServerRecordingRow from './server-recordings/ServerRecordingRow';
 import useRecordingsReSync from '../../hooks/queries/recordings/useRecordingsReSync';
 import Pagy from '../shared/Pagy';
