+++ app/javascript/components/admin/forms/DeleteUserForm.jsx
@@ -6,7 +6,7 @@ import {
 import PropTypes from 'prop-types';
 import Form
   from '../../forms/Form';
-import useDeleteUser from '../../../hooks/mutations/admins/useDeleteUser';
+import useDeleteUser from '../../../hooks/mutations/admin/manage_users/useDeleteUser';
 import Spinner from '../../shared/stylings/Spinner';
 
 export default function DeleteUserForm({ user, handleClose }) {
