+++ app/javascript/hooks/mutations/admin/manage_users/useAdminCreateUser.jsx
@@ -1,6 +1,6 @@
 import { useMutation, useQueryClient } from 'react-query';
 import { toast } from 'react-hot-toast';
-import axios from '../../../helpers/Axios';
+import axios from '../../../../helpers/Axios';
 
 export default function useAdminCreateUser({ onSettled }) {
   const queryClient = useQueryClient();
