+++ app/javascript/hooks/mutations/admin/manage_users/useDeleteUser.jsx
@@ -1,6 +1,6 @@
 import { useMutation, useQueryClient } from 'react-query';
 import { toast } from 'react-hot-toast';
-import axios from '../../../helpers/Axios';
+import axios from '../../../../helpers/Axios';
 
 export default function useDeleteUser(userId) {
   const queryClient = useQueryClient();
