+++ app/javascript/hooks/mutations/admin/room_configuration/useUpdateRoomConfig.jsx
