+++ app/javascript/hooks/mutations/admin/server_rooms/useDeleteServerRoom.jsx
