+++ app/javascript/hooks/queries/admin/manage_users/useActiveUsers.jsx
@@ -1,5 +1,5 @@
 import { useQuery } from 'react-query';
-import axios from '../../../helpers/Axios';
+import axios from '../../../../helpers/Axios';
 
 export default function useActiveUsers(input, page) {
   const params = {
