+++ app/javascript/hooks/queries/admin/room_configuration/useRoomConfigs.jsx
